phase-3/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jose>=1.0.0",
    "openai-agents>=0.6.7",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]
